refactor: introduce `snow/engine/avalanche/vertex/vertextest`
